+++ experimental/benchmarks/web_tooling_benchmark/postdata.py
@@ -30,12 +30,12 @@ if __name__ == '__main__':
         line = f.readline()
         while line:
              if "Geometric mean" in line:
-                  val = int(float(line.split()[-2])*100)
+                  val = float(line.split()[-2])*100
                   print val
                   arr.append(val)
              line = f.readline()
         f.close()
-        res = np.median(arr)
+        res = int(np.median(arr))
         if not res:
              print "no data, post data failed!"
         else:
